==> 2. Code/0_Setup.R <==
options(dplyr.summarise.inform = FALSE)
library(delver)
library(sf)
library(FSA)
library(nlstools)
library(lubridate)
library(stringr)
library(data.table)
library(readr)
library(tidyverse)
library(rlist)
library(Hmisc)
library(ggplot2)
library(survival)
library(ggfortify)
library(gam)
library(cder)
#devtools::install_github("o1iv3r/FeatureImpCluster")
library(FeatureImpCluster)
#devtools::install_github("EnvironmentalScienceAssociates/esaRmisc")
library(esaRmisc)

token = read_file(file.path("1. Data","Inputs","DelveToken.txt"))

delver::set_user_token(token)

source(file.path("2. Code", "functions.R"))
tz_loc = "America/Los_Angeles"
